+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/product/ProductRepository.kt
@@ -3,6 +3,8 @@ package com.app.arcabyolimpo.domain.repository.product
 import com.app.arcabyolimpo.domain.model.product.ProductAdd
 import com.app.arcabyolimpo.domain.model.product.Product
 import kotlinx.coroutines.flow.Flow
+import com.app.arcabyolimpo.domain.model.product.ProductDetail
+import com.app.arcabyolimpo.domain.model.product.ProductUpdate
 
 /**
  * Interface defining the contract for managing product-related data operations.
@@ -28,5 +30,23 @@ interface ProductRepository {
 
     suspend fun searchProducts(query: String): List<Product>
 
+    /**
+     * Retrieves detailed information for a specific product by its ID.
+     *
+     * @param id The unique identifier of the product to retrieve.
+     * @return A [Result] containing the [Product] domain model if successful.
+     */
+    suspend fun getProduct(id: String): Result<ProductDetail>
 
+    /**
+     * Updates an existing product's details and optionally its image.
+     *
+     * @param idProduct The unique identifier of the product to update (passed in the URL path).
+     * @param product The [ProductUpdate] object containing the new details.
+     * @return A [Result] indicating whether the operation was successful or if an error occurred.
+     */
+    suspend fun updateProduct(
+        idProduct: String,
+        product: ProductUpdate
+    ): Result<Unit>
 }